Struct pallet_operator_trading::Module

pub struct Module<T: Trait>(_);

The module declaration.

Methods

impl<T: Trait> Module<T>

Can also be called using Call.

pub fn offer(
origin: T::Origin,
sender: T::AccountId,
contracts: Vec<T::AccountId>,
amount: BalanceOf<T>,
expired: T::BlockNumber
) -> DispatchResult

Offer is an easy contract to trade. If the sender accept during the period, the operator trading will be completed. After the offer, the part of the amount of the buyer's balances will lock.

Note: Only one offer can be issued at the same time each an account. TODO: weight

pub fn reject(origin: T::Origin, offer_id: T::AccountId) -> DispatchResult

Reject the target offer. the offer's buyer or sender can reject the offer. After the reject, the buyer's balances will be unlock. TODO: weight

pub fn accept(origin: T::Origin, offer_id: T::AccountId) -> DispatchResult

Accept the target offer. Only the offer's sender can accept the offer. After the accept:

the buyer's balances will be unlock.

the buyer's balances tranfer to the sender.

the sender's target contracts transfer to the buyer. TODO: weight

pub fn remove(origin: T::Origin) -> DispatchResult

Remove the offer. The offer's owner can remove the offer. But, if the offer is living(until expired), he can not remove the offer. TODO: weight